+++ finalPro/src/main/java/com/kh/spring/lecture/model/service/LectureService.java
@@ -9,13 +9,5 @@ import com.kh.spring.lecture.model.vo.Teacher;
 public interface LectureService {
 	public ArrayList<Lecture> lectureList(Lecture l);
 	public Lecture lectureDetail(int lecNo);
-	//1.게시판 리스트 페이지
-		int selectListCount_Lecture();
-		int selectListCount_Teacher();
-		ArrayList<Lecture> selectLectureList(PageInfo pi);
-		ArrayList<Teacher> selectTeacherList(PageInfo pi);
-			
-		//2.게시글 작성
-		int insertLecture(Lecture l);
-		int insertTeacher(Teacher t);
+	
 }
