+++ finalPro/src/main/resources/mappers/admin-mapper.xml
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "mybatis-3-mapper.dtd" >
+<mapper namespace="adminMapper">
+	
+	<insert id="insertLecture" >
+		insert into lecture(lec_no, lec_name, lec_sdate, lec_edate,
+							 lec_info, lec_stime, lec_etime, lec_count, lec_filename,
+							  lec_bcatg, lec_scatg, lec_price, lec_grade, lec_jnum, lec_cnum,
+							  tea_no, loc_no)
+			values((seq_lno.nextval), #{lecName}, #{lecSdate}, #{lecEdate}, 
+			#{lecInfo}, #{lecStime}, #{lecEtime}, #{lecCount}, #{lecFilename}, 
+			#{lecBcatg}, #{lecScatg}, #{lecPrice}, #{lecGrade}, #{lecJnum}, #{lecCnum},
+			#{teaNo}, #{locNo})
+ 	</insert>
+  	
+  	<select id="selectLectureList" resultMap="lectureMapper.lectureResult">
+		select l.lec_no, l.lec_name, t.tea_name, l.lec_sdate, l.lec_edate, l.lec_stime, l.lec_etime, l.lec_count,
+						 l.lec_bcatg, l.lec_scatg, l.lec_price, l.lec_grade, l.lec_cnum, ll.loc_name
+			from lecture l
+			join Teacher t on (l.tea_no = t.tea_no)
+			join LectureLocation ll on (l.loc_no = ll.loc_no)
+		order by lec_no desc
+	</select>
+	
+	<select id="selectListCount_Lecture" resultType="_int">
+		select count(*)
+			from lecture
+	</select>
+
+	<insert id="insertTeacher">
+		insert into teacher (tea_no, tea_name, tea_age, tea_info, tea_edate)
+		values((seq_tno.nextval), #{teaName}, #{teaAge}, #{teaInfo}, sysdate)
+	</insert>
+	
+  	<select id="selectTeacherList" resultMap="lectureMapper.Teacher">
+		select tea_no, tea_name, tea_age, TO_CHAR(tea_edate,'YY.MM.DD') as "tea_edate", tea_status
+			from teacher
+		order by tea_no desc
+	</select>
+  
+  	<select id="selectListCount_Teacher" resultType="_int">
+		select count(*)
+			from teacher
+		where tea_status = 'Y'
+	</select>
+	
+</mapper>
